Fix notification pop-through.
Because the z height between notifications was set to 0, certain
scenarios could cause the low priority notification to jump in
front of the other notifications in the stack.

Unfortunately, the design of the notifications will suffer from
fixing this, as there will now be shadows when collapsing the
stack. Bummer.

Ticket: CYNGNOS-3010

Change-Id: I7a7e65cb35d2141e542a2de7f51a21960900e8d9