Change structure for bacon specific appfilter internal overlay.
When building Hexo for bacon, we should block the Google apps from
theming automatically. But for all other scenarios, including store,
we can allow those to be themed. Instead of maintaining a separate
branch for the store version, this change allows us to build for the
store and devices from the same base.

Change-Id: Ia8df06db03794000503b8e544ee6f967322b130b